Re-revised workflow in README
I realized mid way through the last workflow that scientific calculators
will take the entire string of inputs and process them in order of
precedence, which I thought only happened when the user explicitly puts
certain operations inside of parenthesis.

With that being the case, I am going to stick with the original project
and just make a regular calculator.