Condense multiple button classes into modifier
I just discovered that the percent, reciprocal, squared and square root
buttons are actually modifiers to the current number, for the most part.

I swapped the square root and clear buttons because it makes more sense
to group the modifier buttons together instead of how Microsoft decided
to lay them out.

Now I need to rework all of them.